1.Create GStreamer pipeline command for i.MX8 platform 2.Temporarily ignore Wi-Fi and temperature retrieval issues on i.MX8 3.Remove self.grabMouse() on i.MX8 to avoid warning messages

+++ ui/ui_sys_sw_info.py
@@ -108,19 +108,21 @@ class UiSystemSoftware(QWidget):
         self.frame_mem_usage_circular_progress.setValue(self.ram_percent)
 
         if "aarch64" in self.machine_arch:
-            temp_str = os.popen("vcgencmd measure_temp").read()
-            self.label_temperature.setText(temp_str)
-            self.temperature_now = temp_str.strip().split("=")[1].split(".")[0]
-            if int(self.temperature_now) > int(self.max_temperature):
-                self.max_temperature_time = datetime.now()
-                self.max_temperature = self.temperature_now
-                # write temp log
-                with open(self.temperature_log_file_uri, "w") as f:
-                    data = "max_temp:{};time:{}".format(self.max_temperature, self.max_temperature_time)
-                    f.write(data)
-                    f.truncate()
-                    f.close()
-                    os.system('sync')
+            if "pi5" in platform.node() or "pi4" in platform.node():
+                temp_str = os.popen("vcgencmd measure_temp").read()
+                self.label_temperature.setText(temp_str)
+                self.temperature_now = temp_str.strip().split("=")[1].split(".")[0]
+                if int(self.temperature_now) > int(self.max_temperature):
+                    self.max_temperature_time = datetime.now()
+                    self.max_temperature = self.temperature_now
+                    # write temp log
+                    with open(self.temperature_log_file_uri, "w") as f:
+                        data = "max_temp:{};time:{}".format(self.max_temperature, self.max_temperature_time)
+                        f.write(data)
+                        f.truncate()
+                        f.close()
+                        os.system('sync')
+            return
 
         # self.temperature_log.debug("max_temperature: %d", 88)
